update program 1
This program is based on the Nick Montfort's Upstart program. I changed the Upstart program to be a wardrobe generator choosing randomized outfits for the user. The program I created here is similar, however the outfits are already chosen for you. All you have to do is type in the occasion and it spits out what you should wear and how many items are in the outfit. This is a good initial start because it combines two functions that is also customizable for the creator too simplify their lives.